requirements.txt
vllm
sentence-transformers
faiss-cpu
fastapi
uvicorn[standard]
pydantic
python-multipart
transformers
huggingface-hub